Use 48 bit [User/Chat/Channel]Id, 56 bit PeerId.

+++ Telegram/SourceFiles/ui/chat/group_call_userpics.h
@@ -18,7 +18,7 @@ namespace Ui {
 struct GroupCallUser {
 	QImage userpic;
 	std::pair<uint64, uint64> userpicKey = {};
-	int32 id = 0;
+	uint64 id = 0;
 	bool speaking = false;
 };
